Moved button-behavior
